requirements.txt
python-telegram-bot==21.9
pexpect==4.9.0
python-dotenv==1.0.1
